backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
tqdm
beautifulsoup4==4.7.1
boto3==1.9.147
requests==2.21.0
Flask==1.0.2
lxml
python-magic==0.4.18
latex2mathml==2.16.2
itsdangerous==2.0.1
python-multipart
mangum
langchain
tiktoken
nltk